api/requirements.txt
elasticsearch==8.5.0
fastapi==0.86.0
graphql-core==3.2.3
nest-asyncio==1.5.6
psutil==5.9.3
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
strawberry-graphql==0.139.0
uvicorn==0.19.0